moved route logics to controller file
